+++ bin/test.sh
@@ -3,12 +3,12 @@
 #
 # SPDX-FileCopyrightText: 2023-2026 JWP Consulting GK
 # Note:
-# Run within uv shell, nix flake shell, or similar environment
+# This script prefixes commands with `uv run`
 set -e
 target="${1-.}"
 echo "Testing $target"
 
-if mypy "$target"
+if uv run mypy "$target"
 then
     echo "mypy ran successfully"
 else
@@ -16,33 +16,33 @@ else
     exit 1
 fi
 
-if ruff format "$target"; then
+if uv run ruff format "$target"; then
     echo "ruff format ran successfully"
 else
     echo "ruff format didn't finish successfully"
     exit 1
 fi
 
-if djlint --check "$target"
+if uv run djlint --check "$target"
 then
     echo "djlint --check ran successfully"
 else
     echo "There was an error running djlint --check"
-    djlint --reformat "$target"
-    if ! djlint --check "$target"; then
+    uv run djlint --reformat "$target"
+    if ! uv run djlint --check "$target"; then
         echo "Couldn't fix with djlint --reformat"
         exit 1
     fi
 fi
 
-if ruff check --fix "$target"; then
+if uv run ruff check --fix "$target"; then
     echo "ruff ran successfully"
 else
     echo "ruff didn't finish successfully"
     exit 1
 fi
 
-if djlint --lint "$target"
+if uv run djlint --lint "$target"
 then
     echo "djlint --lint ran successfully"
 else
@@ -50,7 +50,7 @@ else
     exit 1
 fi
 
-if pytest -n auto --maxprocesses=8 --lf "$target"
+if uv run pytest -n auto --lf "$target"
 then
     echo "pytest ran successfully"
 else
@@ -58,7 +58,7 @@ else
     exit 1
 fi
 
-if vulture .
+if uv run vulture .
 then
     echo "Vulture did not find any issues"
 else
@@ -66,7 +66,7 @@ else
     exit 1
 fi
 
-if reuse lint
+if uv run reuse lint
 then
     echo "reuse lint did not find any issues"
 else