+++ projectify/rules.py
@@ -96,8 +96,6 @@ def can_create_more(
 within_task_quota = rules.predicate(partial(can_create_more, "Task"))
 # Return True if a project can be created in workspace
 within_project_quota = rules.predicate(partial(can_create_more, "Project"))
-# Return True if a section can be created in a workspace
-within_section_quota = rules.predicate(partial(can_create_more, "Section"))
 # Return True if a team member can be added to a workspace
 # The two following use the same quota
 within_team_member_quota = rules.predicate(
@@ -153,14 +151,6 @@ rules.add_perm(
 rules.add_perm("workspace.read_project", is_at_least_observer)
 rules.add_perm("workspace.update_project", is_at_least_maintainer)
 rules.add_perm("workspace.delete_project", is_at_least_maintainer)
-
-# Section
-rules.add_perm(
-    "workspace.create_section", is_at_least_maintainer & within_section_quota
-)
-rules.add_perm("workspace.read_section", is_at_least_observer)
-rules.add_perm("workspace.update_section", is_at_least_maintainer)
-rules.add_perm("workspace.delete_section", is_at_least_maintainer)
 
 # Task
 rules.add_perm(